requirements.txt
requests
beautifulsoup4
pymupdf
faiss-cpu
numpy
sentence-transformers
langchain-openai
pytesseract
